+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/query/OnArtworksQuery.kt
@@ -11,7 +11,7 @@ import android.os.Build
 import android.util.Size
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.OnAudioQueryPlugin
+import com.lucasjosino.on_audio_query.controller.PermissionController
 import com.lucasjosino.on_audio_query.query.helper.OnAudioHelper
 import com.lucasjosino.on_audio_query.types.checkArtworkFormat
 import com.lucasjosino.on_audio_query.types.checkArtworkType
@@ -69,25 +69,30 @@ class OnArtworksQuery : ViewModel() {
         // Define the [type]:
         type = call.argument<Int>("type")!!
 
+        // Request permission status;
+        val hasPermission: Boolean = PermissionController().permissionStatus(context)
+
+        // We cannot 'query' without permission so, throw a PlatformException.
+        if (!hasPermission) {
+            result.error(
+                "403",
+                "The app doesn't have permission to read files.",
+                "Call the [permissionsRequest] method or install a external plugin to handle the app permission."
+            )
+            return
+        }
+
         // Query everything in background for a better performance.
         viewModelScope.launch {
-            // Request permission status from the main method.
-            val hasPermission = OnAudioQueryPlugin().onPermissionStatus(context)
-            // Empty array.
-            var resultArtList: ByteArray? = null
-
-            // We cannot "query" without permission so, just return null.
-            if (hasPermission) {
-                // Start querying
-                resultArtList = loadArt()
-            }
+            // Start querying
+            var resultArtList: ByteArray? = loadArt()
 
             // Sometimes android will extract a 'wrong' or 'empty' artwork. Just set as null.
             if (resultArtList != null && resultArtList.isEmpty()) {
                 resultArtList = null
             }
 
-            // Flutter UI will start, but, information still loading
+            // After loading the information, send the 'result'.
             result.success(resultArtList)
         }
     }