+++ on_audio_query/android/src/main/kotlin/com/lucasjosino/on_audio_query/query/OnAudiosQuery.kt
@@ -6,7 +6,7 @@ import android.content.Context
 import android.net.Uri
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lucasjosino.on_audio_query.OnAudioQueryPlugin
+import com.lucasjosino.on_audio_query.controller.PermissionController
 import com.lucasjosino.on_audio_query.query.helper.OnAudioHelper
 import com.lucasjosino.on_audio_query.types.checkAudiosUriType
 import com.lucasjosino.on_audio_query.types.sorttypes.checkSongSortType
@@ -31,7 +31,6 @@ class OnAudiosQuery : ViewModel() {
 
     @SuppressLint("StaticFieldLeak")
     private lateinit var context: Context
-
 
     /**
      * Method to "query" all songs.
@@ -66,21 +65,26 @@ class OnAudiosQuery : ViewModel() {
             selection = projection[0] + " like " + "'%" + call.argument<String>("path") + "/%'"
         }
 
+        // Request permission status;
+        val hasPermission: Boolean = PermissionController().permissionStatus(context)
+
+        // We cannot 'query' without permission so, throw a PlatformException.
+        if (!hasPermission) {
+            result.error(
+                "403",
+                "The app doesn't have permission to read files.",
+                "Call the [permissionsRequest] method or install a external plugin to handle the app permission."
+            )
+            return
+        }
+
         // Query everything in background for a better performance.
         viewModelScope.launch {
-            // Request permission status from the main method.
-            val hasPermission = OnAudioQueryPlugin().onPermissionStatus(context)
-            // Empty list.
-            var resultSongList = ArrayList<MutableMap<String, Any?>>()
+            // Start querying
+            val queryResult: ArrayList<MutableMap<String, Any?>> = loadSongs()
 
-            // We cannot "query" without permission so, just return a empty list.
-            if (hasPermission) {
-                // Start querying
-                resultSongList = loadSongs()
-            }
-
-            //Flutter UI will start, but, information still loading
-            result.success(resultSongList)
+            // After loading the information, send the 'result'.
+            result.success(queryResult)
         }
     }